refactor: rename grammar.pegjs to parser.peggy
